updated file structure
 - fixed four_square function
 - extracted lucas sequence from lucas primality test
 - consolidated some test files
