+++ numth/lucas_sequence/__init__.py
@@ -1,0 +1,1 @@
+from .modular import by_index, double_index, lucas_sequence_gen
