+++ numth/types/__init__.py
@@ -1,0 +1,4 @@
+from .polynomial import polyn, Polynomial
+from .quadratic import Quadratic
+from .quaternion import Quaternion
+from .rational import frac, Rational
