+++ numth/types/polynomial.py
@@ -4,7 +4,7 @@ from collections import defaultdict
 from functools import reduce
 import re
 
-from .basic import gcd, lcm, mod_inverse, mod_power
+from ..basic import gcd, lcm, mod_inverse, mod_power
 from .rational import frac, Rational
 #===========================================================
